requirements.txt
azure-functions
requests
python-dotenv
pyodbc
pandas